update dx of `useReverification`
