+++ packages/react/src/hooks/index.ts
@@ -2,7 +2,6 @@ export { useAuth } from './useAuth';
 export { useEmailLink } from './useEmailLink';
 export { useSignIn } from './useSignIn';
 export { useSignUp } from './useSignUp';
-export { __experimental_useReverification } from './useReverification';
 export {
   useClerk,
   useOrganization,
@@ -10,4 +9,5 @@ export {
   useSessionList,
   useUser,
   useSession,
+  __experimental_useReverification,
 } from '@clerk/shared/react';
