+++ packages/shared/src/react/hooks/index.ts
@@ -7,3 +7,4 @@ export { useSessionList } from './useSessionList';
 export { useUser } from './useUser';
 export { useClerk } from './useClerk';
 export { useDeepEqualMemo, isDeeplyEqual } from './useDeepEqualMemo';
+export { __experimental_useReverification } from './useReverification';
